fix bug in export command about invalid file path; Update docs

+++ src/main/java/seedu/address/logic/commands/ExportTxtCommand.java
@@ -3,11 +3,13 @@ package seedu.address.logic.commands;
 import static java.util.Objects.requireNonNull;
 
 import java.io.IOException;
+import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.Paths;
 
 import seedu.address.commons.util.XmlToTxtUtil;
 import seedu.address.logic.CommandHistory;
+import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.model.Model;
 import seedu.address.storage.Storage;
 
@@ -17,22 +19,32 @@ import seedu.address.storage.Storage;
  */
 public class ExportTxtCommand extends ExportCommand {
 
-    protected Path exportedFilePath;
+    private String exportedFilePath;
+    private FileType fileType;
 
-    public ExportTxtCommand(Path filePath) {
+    public ExportTxtCommand(String filePath, FileType fileType) {
         super(filePath);
         this.exportedFilePath = filePath;
+        this.fileType = fileType;
     }
 
     @Override
-    public CommandResult execute(Model model, CommandHistory history) {
+    public CommandResult execute(Model model, CommandHistory history) throws CommandException {
         requireNonNull(model);
         requireNonNull(storage);
+
+        if (!isValidTxtFile()) {
+            throw new CommandException(String.format(MESSAGE_INVALID_FILE_PATH));
+        }
+
+        if (!Files.isWritable(Paths.get(exportedFilePath))) {
+            throw new CommandException(String.format(MESSAGE_FILE_PERMISSION_DENIED));
+        }
 
         try {
             Path tempPath = Paths.get("temp.xml");
             storage.saveAddressBook(model.getAddressBook(), tempPath);
-            XmlToTxtUtil.parse(tempPath.toFile(), exportedFilePath.toString());
+            XmlToTxtUtil.parse(tempPath.toFile(), exportedFilePath);
             tempPath.toFile().delete();
         } catch (IOException e) {
             return new CommandResult(String.format(MESSAGE_FAIL_READ_FILE));
@@ -45,6 +57,10 @@ public class ExportTxtCommand extends ExportCommand {
 
     public void setStorage(Storage storage) {
         this.storage = storage;
+    }
+
+    private boolean isValidTxtFile() {
+        return isValidFilePath() && exportedFilePath.endsWith(".txt");
     }
 
     @Override
